+++ BALSAMIC/commands/options.py
@@ -1,4 +1,5 @@
 """Balsamic command options."""
+
 import click
 
 from BALSAMIC import __version__ as balsamic_version
@@ -22,7 +23,11 @@ from BALSAMIC.constants.cluster import (
 from BALSAMIC.constants.constants import LOG_LEVELS, LogLevel
 from BALSAMIC.constants.rules import DELIVERY_RULES
 from BALSAMIC.constants.workflow_params import VCF_DICT
-from BALSAMIC.utils.cli import validate_cache_version, validate_exome_option
+from BALSAMIC.utils.cli import (
+    validate_cache_version,
+    validate_exome_option,
+    validate_umi_min_reads,
+)
 
 OPTION_ANALYSIS_DIR = click.option(
     "--analysis-dir",
@@ -391,3 +396,29 @@ OPTION_TUMOR_SAMPLE_NAME = click.option(
     type=click.STRING,
     help="Tumor sample name",
 )
+<<<<<<< HEAD
+=======
+
+OPTION_UMI = click.option(
+    "--umi/--no-umi",
+    default=True,
+    show_default=True,
+    is_flag=True,
+    help="UMI processing steps for samples with UMI tags. For WGS cases, UMI is always disabled.",
+)
+
+OPTION_UMI_TRIM_LENGTH = click.option(
+    "--umi-trim-length",
+    default=5,
+    show_default=True,
+    type=click.INT,
+    help="Trim N bases from reads in FASTQ file",
+)
+
+OPTION_UMI_MIN_READS = click.option(
+    "--umi-min-reads",
+    type=click.STRING,
+    callback=validate_umi_min_reads,
+    help="Minimum raw reads supporting each UMI group. Format: 'x,y,z'.",
+)
+>>>>>>> develop