+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Fixed:
 * Corrected tool name in deduplication metrics https://github.com/Clinical-Genomics/BALSAMIC/pull/1441
 * MSI table https://github.com/Clinical-Genomics/BALSAMIC/pull/1459
 * Pin numpy version in CNVkit container https://github.com/Clinical-Genomics/BALSAMIC/pull/1457
+* CNVkit incorrect version in the documentation https://github.com/Clinical-Genomics/BALSAMIC/pull/1457
 
 [15.0.1]
 --------